@@ -1,17 +1,26 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import React from "react";
+import ReactDOM from 'react-dom/client';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { Provider } from 'react-redux';
+import { Home } from './pages/home';
+import reportWebVitals from "./reportWebVitals";
+import store from "./store";
 
-import App from './App';
-import store from './app/store';
-
-import 'antd/dist/antd.css';
-ReactDOM.render(
-    <Router>
-        <Provider store={store}>
-            <App />
-        </Provider>
-    </Router>,
-    document.getElementById('root')
+const root = ReactDOM.createRoot(
+    document.getElementById('root') as HTMLElement
 );
+
+root.render(
+    <React.StrictMode>
+        <Router>
+            <Provider store={store}>
+                <Home />
+            </Provider>
+        </Router>
+    </React.StrictMode>
+);
+
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+reportWebVitals();
