@@ -1,10 +1,7 @@
-import React from 'react';
 import { Spin } from 'antd';
 
-const Loader: React.FC = () => (
+export const Loader = () => (
     <div className="loader">
         <Spin />
     </div>
 );
-
-export default Loader;
