@@ -1,5 +1,3 @@
-export { default as Navbar } from './Navbar';
-export { default as Homepage } from './Homepage';
-export { default as Cryptocurrencies } from './Cryptocurrencies';
-export { default as CryptoDetails } from './CryptoDetails';
-export { default as News } from './News';
+export * from './line-chart';
+export * from './loader';
+export * from './navbar';
