@@ -1,25 +1,23 @@
-import React, { Fragment, useState, useEffect } from 'react'
+import { useState, useEffect } from 'react'
 import { Link } from 'react-router-dom';
 import { Card, Row, Col, Input } from 'antd';
-
+import { useGetCryptoCoinsQuery } from '../../api/crypto-api';
+import { CryptoCoin } from '../../models/CryptoCoin';
+import { Loader } from '../../components';
 import millify from 'millify';
-import Loader from '../Loader';
 
-import { useGetCryptoCoinsQuery } from '../../services/crypto-api';
-import { CryptoCoin } from '../../types/CryptoCoin';
-
-interface IProps {
+interface Props {
     simplified?: boolean
 }
 
-const Cryptocurrencies: React.FC<IProps> = props => {
+export const CryptoCurrencies = (props: Props) => {
     const count = props.simplified ? 10 : 100;
     const { data: cryptosList, isFetching } = useGetCryptoCoinsQuery(count);
     const [cryptos, setCryptos] = useState<CryptoCoin[]>([]);
     const [searchTerm, setSearchTerm] = useState<string>('');
 
     useEffect(() => {
-        let coins = cryptosList?.data?.coins;
+        let coins = cryptosList?.coins;
         if (coins) {
             const filteredData = coins.filter((coin) => coin.name.toLowerCase().includes(searchTerm.toLocaleLowerCase()));
             if (filteredData) setCryptos(filteredData);
@@ -29,7 +27,7 @@ const Cryptocurrencies: React.FC<IProps> = props => {
     if (isFetching) return <Loader />;
 
     return (
-        <Fragment>
+        <>
             {!props.simplified && (
                 <div className='search-crypto'>
                     <Input placeholder='Search Cryptocurrency' onChange={(e) => setSearchTerm(e.target.value)} />
@@ -53,8 +51,6 @@ const Cryptocurrencies: React.FC<IProps> = props => {
                     </Col>
                 ))}
             </Row>
-        </Fragment>
+        </>
     )
 }
-
-export default Cryptocurrencies;
