@@ -1,21 +1,21 @@
-import React, { Fragment } from 'react';
 import { Link } from 'react-router-dom';
 import { Typography, Row, Col, Statistic } from 'antd';
-
+import { useGetCryptoCoinsQuery } from '../../../api/crypto-api';
+import { Loader } from '../../../components/loader';
+import { CryptoCurrencies } from '../../crypto';
+import { News } from '../../news';
 import millify from 'millify';
-import Loader from '../Loader';
-
-import { Cryptocurrencies, News } from '../../components';
-import { useGetCryptoCoinsQuery } from '../../services/crypto-api';
 
 const { Title } = Typography;
 
-const Homepage: React.FC = () => {
+export const HomeDetails = () => {
   const { data, isFetching } = useGetCryptoCoinsQuery(10);
 
-  if (isFetching) return <Loader />;
+  if (isFetching) {
+    return <Loader />;
+  }
 
-  const globalStats = data?.data?.stats;
+  const globalStats = data?.stats;
   const totalCryptocurrencies = globalStats?.total ? globalStats?.total : 0;
   const totalExchanges = globalStats?.totalExchanges ? globalStats?.totalExchanges : 0;
   const totalMarkets = globalStats?.totalMarkets ? globalStats?.totalMarkets : 0;
@@ -24,7 +24,7 @@ const Homepage: React.FC = () => {
   const total24hPercent = (total24hVolume / totalMarketCap) * 100;
 
   return (
-    <Fragment>
+    <>
       <Title level={2} className='heading'>Global Crypto Stats</Title>
       <Row gutter={[24, 24]}>
         <Col xs={24} sm={12} lg={8}>
@@ -42,14 +42,12 @@ const Homepage: React.FC = () => {
         <Title level={2} className='home-title'>Top 10 Cryptocurrencies in the world</Title>
         <Title level={3} className='show-more'><Link to='cryptocurrencies'>Show More</Link></Title>
       </div>
-      <Cryptocurrencies simplified={true} />
+      <CryptoCurrencies simplified={true} />
       <div className='home-heading-container'>
         <Title level={2} className='home-title'>Latest Crypto News</Title>
         <Title level={3} className='show-more'><Link to='news'>Show More</Link></Title>
       </div>
       <News simplified={true} />
-    </Fragment>
+    </>
   )
 }
-
-export default Homepage;
